tdk_robokit1: Use 24MHz SPI for everything
It turns out even with all the other SPI fixes once the clock is set
in the SPI_CSR register, updating it does not seem to actually update the
clock. This is likely due to the impossible to solve issue... where the
peripheral does not know when to clock/unclock with gpio chip selects
seemingly.

Signed-off-by: Tom Burdick <thomas.burdick@intel.com>