soc: ti: k3: am6x: m4: Adjust default RAT configuration
The current default RAT configuration puts the first 512MB of
address space on top of the first 512MB of DRAM. This DRAM
can include the carveout region assigned for IPC with the
host core, for instance the M4 on AM62 uses 0x9cc00000.

Split this first translation from 0x80000000 to 0x60000000 and
0xc0000000. Shrink the 0x40000000 translation and move down
to 0x70000000. This provides the same system IP coverage without
blocking access to the first 512MB of DRAM.

While here, fix typo with leading 0 for a couple BASE_ADDR
translations, no functional change.

Signed-off-by: Andrew Davis <afd@ti.com>